+++ modelwhaleutils.egg-info/requires.txt
@@ -3,3 +3,4 @@ requests
 pynvml
 psutil
 boto3
+mlflow
